Implement half of an automated regression test suite
Relevant to #333

The half implemented here is for managing a queue of jobs based on the
current GitHub state of the repositories. The half remaining to be
implemented is workers that will interact with this queue and actually
run regression tests.

Given that barebones setup, we could next consider adding a nicer web
interface to the worldwide state of regression tests. (A very basic
view is available via the HTTP API already.)