Add ConfigGC op to closLang and below
Currently, this operation forces a GC to run. The semantics of this
new op is to do nothing noticeable because abstract languages can't
see the GC. The op is implemented in data_to_word, i.e. it doesn't
exist as an op in languages below dataLang.

The new op takes two arguments. Currently the arguments are ignored,
but later these will be used to determine what ConfigGC is to do. For
example, some combination of inputs will focre a full GC. Another
combination of inputs will switch from the simple GC to the
generational GC (with a nursery of a given size). Knowing what
combination of does what is not user friendly, so later there will be
functions in the basis library which will hide the arcane numeric
combinations from the user. Nothing should parse to ConfigGC.