Make sure that addancs gets an old time-stamp when reloc-building
This means that we can accurately record a dependency for the script
files on the corresponding lem files and the addancs tool, but the
latter will not cause a rebuild even though it will have just been
rebuilt. This is the sane as what is being done inside Holmake with
"heap" files.

Requires HOL-Theorem-Prover/HOL@514a40b4a5 or later.